pyproject.toml
[project]
name = "identity-photo-editor"
version = "0.1.0"
description = "Identity-preserved photo editing system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
  "prometheus-client>=0.20.0",
  "celery>=5.4.0",
  "redis>=5.0.0",
  "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
